Move _setup to setup.py for extension
